requirements.txt
streamlit
pandas
pdfplumber
requests
google-genai
python-dotenv